+++ src/cucumber.rs
@@ -997,7 +997,7 @@ where
             &'a gherkin::Feature,
             Option<&'a gherkin::Rule>,
             &'a gherkin::Scenario,
-            Option<&'a W>,
+            Option<&'a mut W>,
         ) -> LocalBoxFuture<'a, ()>
         + 'static,
 {
@@ -1108,7 +1108,7 @@ where
                 &'a gherkin::Feature,
                 Option<&'a gherkin::Rule>,
                 &'a gherkin::Scenario,
-                Option<&'a W>,
+                Option<&'a mut W>,
             ) -> LocalBoxFuture<'a, ()>
             + 'static,
     {